// frontend/src/app/models/User.ts
export interface User {
  email: string;
  password: string;
  name: string;
  phone: string;
  street: string;
  street_number: number;
  district: string;
  additional_address_details: string;
  zip_code: string;
  id_city: number;
  type: number;
  cpf_cnpj: string;
  date_of_birth: string;
  sex: string;
}